refactor(spf): inject VTT parser via provider config, drop dom wrapper
Eliminates the dom/actors/text-track-segment-loader.ts wrapper that
was currying parseVttSegment into the media-level factory. The
provider behavior now declares `config.parseSegment` and receives
the parser through the composition config channel — the pattern
we'd agreed would be right for DOM-bound dependencies.

Changes:

- provideTextTrackActors: signature grows a `config: { parseSegment:
  CueParser<VTTCue> }` requirement. Imports createTextTrackSegmentLoaderActor
  from media/actors/ directly (no wrapper) and threads the parser through.

- hls-engine: adds a provideDomTextTrackActors wrapper that closes
  over parseVttSegment, matching the existing pattern for config-
  threading wrappers (selectVideoTrackFromConfig, etc.). Swaps the
  engine's TextTrackSegmentLoaderActor type import to the media/actors
  path (the dom wrapper no longer re-exports it).

- dom/actors/text-track-segment-loader.ts: deleted. Consumers import
  from media/actors directly and pass the parser explicitly.

- Test updates: the actor's direct test imports the factory from
  media/actors and passes parseVttSegment to each call site. The
  load-text-track-cues test supplies parseSegment in provider config.

Net: one fewer file, one less indirection, and the DOM/host seam is
now crisply at the composition-config layer. The parser doesn't leak
into HlsPlaybackEngineConfig — it's closed over by the engine's
internal wrapper, same as how other config-threading wrappers work.

All green: build (160 files), tests (707), typecheck, workspace
check, lint.
